backend/requirements.txt
fastapi
uvicorn
numpy
requests
python-dotenv
rank-bm25
PyPDF2
langchain-text-splitters
tqdm
faiss-cpu==1.13.2
python-multipart
openai
groq